Add generic redeemRewards and redeemerContract functions ; removing redundant code
